part4/requirements.txt
alembic
click
python-dateutil
flask
flask-bcrypt
flask-cors
flask-jwt-extended
flask-login
flask-migrate
flask-restx
flask-sqlalchemy
requests
sqlalchemy
werkzeug
